backend/app/requirements.txt
fastapi==0.115.0
fastapi-jwt[authlib]==0.3.0
uvicorn==0.31.0
gunicorn==21.0.0
python-multipart==0.0.11
requests==2.32.3
# dotenv==0.9.9  <- Replaced by pydantic-settings
pydantic-settings==2.10.1
httpx==0.28.1
xgboost
beartype
python-docx
pdfplumber
pypdf
huggingface_hub
six
opencv-python
onnxruntime
shapely
pyclipper
datrie
hanziconv
nltk
chardet
tiktoken
pandas
openpyxl
python-pptx
readability-lxml
lxml_html_clean
html-text
markdown
elasticsearch
elasticsearch-dsl
xxhash
tika
openai>=1.0.0
jieba
llama-index-core
llama-index-postprocessor-dashscope-rerank-custom
dashscope
psycopg2-binary
bcrypt
fastapi-jwt
python-jose
authlib
colorlog
redis==5.0.1
aioredis==2.0.1
alembic==1.13.1
sqlalchemy==1.4.53
loguru==0.7.3
sentence-transformers==5.0.0
torch==2.5.1